requirements.txt
pandas>=1.3.0
numpy>=1.21.0
tweepy>=4.10.0
requests>=2.25.0
google-api-python-client>=2.80.0
apache-airflow>=2.3.0
python-dotenv>=0.19.0
pyarrow>=6.0.0  # For parquet support